Move game and square components to separate jsx files
