+++ src/components/board/Board.jsx
@@ -1,12 +1,5 @@
 import React from 'react';
-
-function Square(props) {
-    return (
-        <button className="square" onClick={() => props.onClick()}>
-            {props.value}
-        </button>
-    );
-}
+import Square from './square/Square';
 
 export default class Board extends React.Component {
   renderSquare(i) {
