+++ src/components/board/square/Square.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Square extends React.Component {
+    render() {
+        return (
+            <button className="square" onClick={() => this.props.onClick()}>
+                {this.props.value}
+            </button>
+        );
+    }
+}
